change: connections to primary/secondary/routed hosts are not shared
The connections opened when fetching the info necessary to update the members of the replica set, instead, are stored and reused.
Re-use in this case is not a problem, because we have moved inside the 'modifyMVar' the code that fetches the replica info. In this way, only one thread can update the members at a given time, and therefore only one thread can use the stored pipes.