Do not cache connections to replica set members
This commit addresses #119 i.e. the fact that pipes to replica set
members are reused. That choice is not thread-safe. For instance:
- thread 1 asks for a pipe to some host of the replica set; a new pipe is
  opened and returned.
- after that, thread 2 asks for another pipe; updateMembers may reuse the pipe
  returned to thread 1 to fetch the replica info, which is bad.

This commit disables previous pipes to be stored into a ReplicaSet object.
Subsequent calls to updateMembers will open a new pipe to fetch the
replica info. However, that pipe is not closed right away, but
temporarily kept in case the host we want to connect to is the same
as the one that has been used to fetch the replica info. If it isn't,
the pipe is closed and not stored as it was before.